requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile --output-file=requirements.txt pyproject.toml
#
asgiref==3.10.0
    # via
    #   channels
    #   daphne
    #   django
    #   django-micboard (pyproject.toml)
attrs==25.4.0
    # via
    #   service-identity
    #   twisted
autobahn==25.10.1
    # via daphne
automat==25.4.16
    # via twisted
certifi==2025.10.5
    # via
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
channels==4.3.1
    # via django-micboard (pyproject.toml)
charset-normalizer==3.4.4
    # via requests
constantly==23.10.4
    # via twisted
cryptography==46.0.3
    # via
    #   autobahn
    #   pyopenssl
    #   service-identity
daphne==4.2.1
    # via django-micboard (pyproject.toml)
django==5.2.7
    # via
    #   channels
    #   django-micboard (pyproject.toml)
    #   django-picklefield
    #   django-q2
    #   djangorestframework
django-picklefield==3.3
    # via django-q2
django-q-sentry==0.1.6
    # via django-q2
django-q2[sentry]==1.8.0
    # via django-micboard (pyproject.toml)
djangorestframework==3.16.1
    # via django-micboard (pyproject.toml)
hyperlink==21.0.0
    # via
    #   autobahn
    #   twisted
idna==3.11
    # via
    #   hyperlink
    #   requests
    #   twisted
incremental==24.7.2
    # via twisted
pillow==12.0.0
    # via django-micboard (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   service-identity
pyasn1-modules==0.4.2
    # via service-identity
pycparser==2.23
    # via cffi
pyopenssl==25.3.0
    # via twisted
requests==2.32.5
    # via django-micboard (pyproject.toml)
sentry-sdk==2.42.1
    # via django-q-sentry
service-identity==24.2.0
    # via twisted
sqlparse==0.5.3
    # via django
twisted[tls]==25.5.0
    # via daphne
txaio==25.9.2
    # via autobahn
typing-extensions==4.15.0
    # via twisted
urllib3==2.5.0
    # via
    #   django-micboard (pyproject.toml)
    #   requests
    #   sentry-sdk
websockets==15.0.1
    # via django-micboard (pyproject.toml)
zope-interface==8.0.1
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
